televir - lynt; move entrez wrapper;
