+++ .env_model
@@ -1,3 +1,5 @@
+### Your email for NCBI downloads.
+NCBI_EMAIL=your-email-here
 
 ### change this key, with same length
 SECRET_KEY=23h5vdljj(3mvv28vtwf8)ev_^!$$asnf2t9&hxw97(9j#=lj9
