+++ pathogen_identification/management/deprec_commands/submit_televir_explify_merge.py
@@ -9,7 +9,7 @@ from pathogen_identification.models import (FinalReport, PIProject_Sample,
                                             Projects)
 from pathogen_identification.templatetags.report_colors import \
     flag_false_positive
-from pathogen_identification.utilities.explify_merge import (
+from pathogen_identification.utilities.deprec_utils.explify_merge import (
     get_illumina_found, merge_panels, process_televir, read_panel)
 from pathogen_identification.utilities.utilities_general import get_project_dir
 from utils.process_SGE import ProcessSGE
