+++ pathogen_identification/management/deprec_commands/submit_televir_explify_merge_ext.py
@@ -5,7 +5,7 @@ from django.contrib.auth.models import User
 from django.core.management.base import BaseCommand
 
 from managing_files.models import ProcessControler
-from pathogen_identification.utilities.explify_merge import (
+from pathogen_identification.utilities.deprec_utils.explify_merge import (
     get_illumina_found, merge_panels, process_televir, read_panel)
 from pathogen_identification.utilities.utilities_general import \
     get_services_dir
